Simplify label address assignment, add test file, and fix program.m
In assembler.c, simplified the assignment of labels->address by directly assigning the value of current_token, assuming it is already correctly initialized. This change was done to clean the code and improve readability.

Also, added a new register test file in programs/reg-test.s. This test was needed to ensure the robustness of the register operations.